// src/domain/enums.ts
export enum VesselType {
  BritishTeapot = 0,
  Zisha = 1,
  WhitePorcelain = 2,
  Celadon = 3,
  JapaneseCeramics = 4,
  Glass = 5,
  GaiwanWhitePorcelain = 6,
  GaiwanCeladon = 7,
}

export enum BoilMethod {
  Electric = 0,
  PotIron = 1,
  PotBronze = 2,
  PotClay = 3,
}

export enum TeaType {
  GreenNormal = 0,
  White = 1,
  Black = 2,
  Oolong = 3,
  Yellow = 4,
  Puerh = 5,
  GreenGyokuro = 6,
  GreenSencha = 7,
  GreenFukamushi = 8,
  Tibetan = 9,
}

export enum UnitSystem {
  Metric = 0,
  US = 1,
  UK = 2,
}

export enum HardnessUnit {
  TdsPpm = 0,
  FH = 1,
  DH = 2,
}

export enum BrewingRegion {
  EastAsia = 'east_asia',
  British = 'british',
  SoutheastAsia = 'southeast_asia',
  Tibetan = 'tibetan_region',
  WesternModern = 'western_modern',
}